+++ downward/experiment.py
@@ -5,6 +5,7 @@ A module for running Fast Downward experiments.
 from collections import defaultdict, OrderedDict
 import logging
 import os.path
+from pathlib import Path
 
 from downward import suites
 from downward.cached_revision import CachedFastDownwardRevision
@@ -18,7 +19,10 @@ from lab.experiment import Experiment, get_default_data_dir, Run
 
 
 class FastDownwardAlgorithm:
-    """Collection of information about a Fast Downward algorithm."""
+    """
+    A Fast Downward algorithm is the combination of revision, driver options and
+    component options.
+    """
 
     def __init__(
         self,
@@ -51,28 +55,25 @@ _DownwardAlgorithm = FastDownwardAlgorithm
 
 
 class FastDownwardRun(Run):
-    def __init__(
-        self, exp: Experiment, algo: FastDownwardAlgorithm, task: suites.Problem
-    ):
+    """An experiment run that uses *algo* to solve *task*.
+
+    See :py:class:`Run <lab.experiment.Run>` for inherited methods.
+
+    """
+
+    def __init__(self, exp: Experiment, algo: FastDownwardAlgorithm, task: suites.Task):
         super().__init__(exp)
-        self.algo = algo
-        self.task = task
+        driver_options = algo.driver_options[:]
 
-        self.driver_options = algo.driver_options[:]
-
-        if self.task.domain_file is None:
-            self.add_resource("task", self.task.problem_file, "task.sas", symlink=True)
+        if task.domain_file is None:
+            self.add_resource("task", task.problem_file, "task.sas", symlink=True)
             input_files = ["{task}"]
             # Without PDDL input files, we can't validate the solution.
-            self.driver_options = [
-                opt for opt in self.driver_options if opt != "--validate"
-            ]
+            driver_options = [opt for opt in driver_options if opt != "--validate"]
         else:
+            self.add_resource("domain", task.domain_file, "domain.pddl", symlink=True)
             self.add_resource(
-                "domain", self.task.domain_file, "domain.pddl", symlink=True
-            )
-            self.add_resource(
-                "problem", self.task.problem_file, "problem.pddl", symlink=True
+                "problem", task.problem_file, "problem.pddl", symlink=True
             )
             input_files = ["{domain}", "{problem}"]
 
@@ -84,27 +85,27 @@ class FastDownwardRun(Run):
             "planner",
             [tools.get_python_executable()]
             + [driver]
-            + self.driver_options
+            + driver_options
             + input_files
             + algo.component_options,
         )
 
-        self._set_properties()
+        self._set_properties(algo, driver_options, task)
 
-    def _set_properties(self):
-        self.set_property("algorithm", self.algo.name)
-        self.set_property("repo", self.algo.cached_revision.repo)
-        self.set_property("local_revision", self.algo.cached_revision.local_rev)
-        self.set_property("global_revision", self.algo.cached_revision.global_rev)
-        self.set_property("build_options", self.algo.cached_revision.build_options)
-        self.set_property("driver_options", self.driver_options)
-        self.set_property("component_options", self.algo.component_options)
+    def _set_properties(self, algo, driver_options, task):
+        self.set_property("algorithm", algo.name)
+        self.set_property("repo", algo.cached_revision.repo)
+        self.set_property("local_revision", algo.cached_revision.local_rev)
+        self.set_property("global_revision", algo.cached_revision.global_rev)
+        self.set_property("build_options", algo.cached_revision.build_options)
+        self.set_property("driver_options", driver_options)
+        self.set_property("component_options", algo.component_options)
 
-        for key, value in self.task.properties.items():
+        for key, value in task.properties.items():
             self.set_property(key, value)
 
         self.set_property("experiment_name", self.experiment.name)
-        self.set_property("id", [self.algo.name, self.task.domain, self.task.problem])
+        self.set_property("id", [algo.name, task.domain, task.problem])
 
 
 class FastDownwardExperiment(Experiment):
@@ -224,8 +225,8 @@ class FastDownwardExperiment(Experiment):
         """
         if isinstance(suite, str):
             suite = [suite]
-        benchmarks_dir = os.path.abspath(benchmarks_dir)
-        if not os.path.exists(benchmarks_dir):
+        benchmarks_dir = Path(benchmarks_dir).resolve()
+        if not benchmarks_dir.is_dir():
             logging.critical(f"Benchmarks directory {benchmarks_dir} not found.")
         self._suites[benchmarks_dir].extend(suite)
 
@@ -342,12 +343,12 @@ class FastDownwardExperiment(Experiment):
         if not self._algorithms:
             logging.critical("You must add at least one algorithm.")
 
-        # We convert the problems in suites to strings to avoid errors when converting
-        # properties to JSON later. The clean but more complex solution would be to add
-        # a method to the JSONEncoder that recognizes and correctly serializes the class
-        # Problem.
+        # We convert the problems in suites to strings to avoid errors when
+        # converting properties to JSON later. The clean but more complex
+        # solution would be to add a method to the JSONEncoder that recognizes
+        # and correctly serializes the Path and Task classes.
         serialized_suites = {
-            benchmarks_dir: [str(problem) for problem in benchmarks]
+            str(benchmarks_dir): [str(problem) for problem in benchmarks]
             for benchmarks_dir, benchmarks in self._suites.items()
         }
         self.set_property("suite", serialized_suites)